Fix timeout issue not being triggered
The timeout isn't being triggered because the last is being set before the timeout time. So to fix I rearranged the logic to make it so that when the state of the light is changed, it changes a variable called `lastSent` this way, when the timeout checks to see if it can be sent, it looks to see if the state of the light is already set to the same or different light state (and same for when turning on the light).